ROX-17857: Show configuration details
